Move addressSearchBar to components folder
