+++ src/components/AddressSearchBar/index.js
@@ -2,8 +2,8 @@ import { withStyles } from '@material-ui/core';
 import { connect } from 'react-redux';
 import { injectIntl } from 'react-intl';
 import AddressSearchBar from './AddressSearchBar';
-import styles from '../../styles';
-import { getLocaleString } from '../../../../redux/selectors/locale';
+import styles from './styles';
+import { getLocaleString } from '../../redux/selectors/locale';
 
 const mapStateToProps = (state) => {
   const { locale } = state.user;
