+++ src/components/AddressSearchBar/styles.js
@@ -1,0 +1,17 @@
+const styles = theme => ({
+  searchBar: {
+    paddingLeft: theme.spacing.unitDouble,
+    marginTop: theme.spacing.unit,
+    border: '1px solid #ACACAC',
+    borderRadius: 4,
+    width: '100%',
+    height: '80%',
+    boxSizing: 'border-box',
+    backgroundColor: '#fff',
+  },
+  clearButton: {
+    fontSize: 22,
+  },
+});
+
+export default styles;
