+++ src/views/AreaView/styles.js
@@ -75,22 +75,9 @@ const styles = theme => ({
   addressTitle: {
     marginBottom: theme.spacing.unitDouble,
   },
-  searchBar: {
-    paddingLeft: theme.spacing.unitDouble,
-    marginTop: theme.spacing.unit,
-    border: '1px solid #ACACAC',
-    borderRadius: 4,
-    width: '100%',
-    height: '80%',
-    boxSizing: 'border-box',
-    backgroundColor: '#fff',
-  },
   fieldFocus: {
     outline: '2px solid transparent',
     boxShadow: `0 0 0 4px ${theme.palette.focusBorder}`,
-  },
-  clearButton: {
-    fontSize: 22,
   },
   addressItem: {
     fontSize: 24,
